refactor(chatMessages): split up into multiple files
