+++ packages/web-main/src/pages/gameserver/cards/OnlinePlayers.tsx
@@ -5,7 +5,7 @@ import { useSelectedGameServer } from 'hooks/useSelectedGameServerContext';
 import { useSocket } from 'hooks/useSocket';
 import { usePlayerOnGameServers, usePlayers } from 'queries/players/queries';
 import { FC, useEffect } from 'react';
-import { Card } from './style';
+import { StyledCard } from './style';
 
 const Container = styled.div`
   display: flex;
@@ -53,7 +53,7 @@ export const OnlinePlayersCard: FC = () => {
   if (isLoading || isLoadingPlayers) return <Skeleton variant="rectangular" width="100%" height="100%" />;
 
   return (
-    <Card variant="outline">
+    <StyledCard variant="outline">
       <Container>
         <h2>{data?.data.length} Players Online</h2>
         <Players>
@@ -65,6 +65,6 @@ export const OnlinePlayersCard: FC = () => {
           })}
         </Players>
       </Container>
-    </Card>
+    </StyledCard>
   );
 };
